[walnut] refactor parsing extraVars
