+++ coconut/control/control.go
@@ -28,8 +28,6 @@ package control
 
 import (
 	"context"
-	"encoding/csv"
-	"encoding/json"
 	"errors"
 	"fmt"
 	"io"
@@ -41,11 +39,11 @@ import (
 	"time"
 
 	"github.com/xlab/treeprint"
-	"gopkg.in/yaml.v3"
 
 	"github.com/AliceO2Group/Control/coconut"
 	"github.com/AliceO2Group/Control/coconut/protos"
 	"github.com/AliceO2Group/Control/common/logger"
+	"github.com/AliceO2Group/Control/common/utils"
 	"github.com/briandowns/spinner"
 	"github.com/olekukonko/tablewriter"
 	"github.com/sirupsen/logrus"
@@ -175,21 +173,6 @@ func GetEnvironments(cxt context.Context, rpc *coconut.RpcClient, cmd *cobra.Com
 	return nil
 }
 
-
-func readAsCSV(val string) ([]string, error) {
-	if val == "" {
-		return []string{}, nil
-	}
-	stringReader := strings.NewReader(val)
-	csvReader := csv.NewReader(stringReader)
-	return csvReader.Read()
-}
-
-func isJson(str string) bool {
-	var js json.RawMessage
-	return json.Unmarshal([]byte(str), &js) == nil
-}
-
 func CreateEnvironment(cxt context.Context, rpc *coconut.RpcClient, cmd *cobra.Command, args []string, o io.Writer) (err error) {
 	wfPath, err := cmd.Flags().GetString("workflow-template")
 	if err != nil {
@@ -212,49 +195,9 @@ func CreateEnvironment(cxt context.Context, rpc *coconut.RpcClient, cmd *cobra.C
 		return
 	}
 
-	extraVarsMap := make(map[string]string)
-
-	if isJson(extraVars) {
-		extraVarsMapI := make(map[string]interface{})
-		err = yaml.Unmarshal([]byte(extraVars), &extraVarsMapI)
-		if err != nil {
-			err = fmt.Errorf("cannot parse extra-vars as JSON: %w", err)
-			return
-		}
-		for k, v := range extraVarsMapI {
-			if strVal, ok := v.(string); ok {
-				extraVarsMap[k] = strVal
-				continue
-			}
-			marshaledValue, marshalErr := json.Marshal(v)
-			if marshalErr != nil {
-				continue
-			}
-			extraVarsMap[k] = string(marshaledValue)
-		}
-	} else {
-		extraVarsSlice := make([]string, 0)
-		extraVarsSlice, err = readAsCSV(extraVars)
-		if err != nil {
-			err = fmt.Errorf("cannot parse extra-vars as CSV: %w", err)
-			return
-		}
-
-		for _, entry := range extraVarsSlice {
-			if len(entry) < 3 { // can't be shorter than a=b
-				err = fmt.Errorf("invalid variable assignment %s", entry)
-				return
-			}
-			if strings.Count(entry, "=") != 1 {
-				err = fmt.Errorf("invalid variable assignment %s", entry)
-				return
-			}
-
-			sanitized := strings.Trim(strings.TrimSpace(entry), "\"'")
-
-			entryKV := strings.Split(sanitized, "=")
-			extraVarsMap[entryKV[0]] = entryKV[1]
-		}
+	extraVarsMap, err := utils.ParseExtraVars(extraVars)
+	if err != nil {
+		return
 	}
 
 	// TODO: add support for setting visibility here OCTRL-178
